Move ini configuration headers to seperate directory
